Add support for APIGatewayV1 (#6)
This commit expands support of the Lambda handler to allow for both Gateway V1
and V2 events. Primarily, there are some slight differences in the shape of the
incoming event between the two versions. The handler is now additionally
responsible for identifying the type of incoming event so that it can transform the
useful bits of data accordingly for Apollo Server.